+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [unreleased]
+
+* No changes yet
+
 ## [0.20.0] - 2020-12-20
 
 ### Added