+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [unreleased]
+
 ## [0.21.0] - 2020-12-20
 
 ### Added